% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{alter}
\alias{alter}
\title{Combine values into a unique vector}
\usage{
alter(...)
}
\arguments{
\item{...}{Vectors to combine.}
}
\value{
A vector of unique values.
}
\description{
Combine values into a unique vector
}
\keyword{internal}
