% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneck_methods.R
\name{cas}
\alias{cas}
\title{Calculate a conditional association score}
\usage{
cas(x, y, z)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}

\item{z}{A numeric conditioning variable or matrix.}
}
\value{
A numeric conditional association score.
}
\description{
Calculate a conditional association score
}
\keyword{internal}
