% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneck_methods.R
\name{cmi}
\alias{cmi}
\title{Calculate conditional mutual information}
\usage{
cmi(v1, v2, nargin = 2, vcs = NULL)
}
\arguments{
\item{v1}{A numeric vector.}

\item{v2}{A numeric vector.}

\item{nargin}{Number of supplied variable groups.}

\item{vcs}{Optional conditioning variable or matrix.}
}
\value{
A numeric conditional mutual information value.
}
\description{
Calculate conditional mutual information
}
\keyword{internal}
