% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{combineAndRmCyc}
\alias{combineAndRmCyc}
\title{Combine DAGs and remove cycles}
\usage{
combineAndRmCyc(DAGs, mode = "single_cell", Emin = NULL, Emax = NULL)
}
\arguments{
\item{DAGs}{A list of learned DAGs.}

\item{mode}{Character string specifying \code{"single_cell"} or \code{"bulk"} mode.}

\item{Emin}{Minimum number of arcs required for a DAG to be retained.}

\item{Emax}{Maximum number of arcs allowed for a DAG to be retained.}
}
\value{
A \code{bn} object representing the combined acyclic graph.
}
\description{
Combine learned DAGs and remove cycles from the resulting network.
}
