% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{combineDAGs}
\alias{combineDAGs}
\title{Combine DAGs}
\usage{
combineDAGs(dags, Emin = NULL, Emax = NULL)
}
\arguments{
\item{dags}{A list of network structures returned by \code{DG_grid()}.}

\item{Emin}{Minimum number of arcs required for a network to be retained.}

\item{Emax}{Maximum number of arcs allowed for a network to be retained.}
}
\value{
A data frame containing the summarized network structure.
}
\description{
Combine network structures after filtering them by their numbers of arcs.
}
