% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{df2bn}
\alias{df2bn}
\title{Convert a graph data frame to a Bayesian network}
\usage{
df2bn(dS, node_names, plot = TRUE)
}
\arguments{
\item{dS}{A data frame with columns named \code{from} and \code{to}.}

\item{node_names}{Character vector of node names.}

\item{plot}{Logical; whether to plot the network.}
}
\value{
A \code{bn} object.
}
\description{
Convert a graph data frame to a Bayesian network
}
