% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{df2mat}
\alias{df2mat}
\title{Convert a network data frame to a matrix}
\usage{
df2mat(DAG, ctypes = NULL)
}
\arguments{
\item{DAG}{A summarized network structure.}

\item{ctypes}{Optional character vector of cell types.}
}
\value{
An adjacency matrix corresponding to the summarized network.
}
\description{
Transform a summarized network structure into an adjacency matrix.
}
