% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findFeatures_func.R
\name{diffScore}
\alias{diffScore}
\title{Get diffScore}
\usage{
diffScore(diffCoeff, edgeSet, abs = TRUE)
}
\arguments{
\item{diffCoeff}{A data frame of previously calculated coefficient
differences.}

\item{edgeSet}{A character vector of edge names.}

\item{abs}{Logical; whether to sum the absolute coefficient differences.}
}
\value{
A numeric vector of differential scores.
}
\description{
Get diffScore
}
