% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectMat_func.R
\name{.effect_energy1d}
\alias{.effect_energy1d}
\title{Calculate the one-dimensional energy distance}
\usage{
.effect_energy1d(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
A non-negative numeric scalar.
}
\description{
Calculate the one-dimensional energy distance
}
\keyword{internal}
