% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectMat_func.R
\name{.effect_mmd_linear}
\alias{.effect_mmd_linear}
\title{Calculate a linear-time MMD estimate}
\usage{
.effect_mmd_linear(x, y)
}
\arguments{
\item{x}{A numeric vector containing the reference values.}

\item{y}{A numeric vector containing the perturbed values.}
}
\value{
A non-negative numeric scalar.
}
\description{
Calculate a linear-time MMD estimate
}
\keyword{internal}
