% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findFeatures_func.R
\name{dsRank}
\alias{dsRank}
\title{Rank genes according to diffScores}
\usage{
dsRank(diffScores)
}
\arguments{
\item{diffScores}{A named numeric vector of previously calculated
differential scores.}
}
\value{
A character vector of ranked gene names.
}
\description{
Rank genes according to diffScores
}
