% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{edgeFilter}
\alias{edgeFilter}
\title{Filter edges by bn coefficients}
\usage{
edgeFilter(dag, data, ctypes, threshold = 0.1)
}
\arguments{
\item{dag}{A \code{bn} object.}

\item{data}{A data frame or matrix used to fit the network.}

\item{ctypes}{Character vector of node names.}

\item{threshold}{Minimum absolute coefficient required to retain an edge.}
}
\value{
A filtered \code{bn} object.
}
\description{
Filter edges by bn coefficients
}
