% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectMat_func.R
\name{g_per}
\alias{g_per}
\title{Permute expression values for a specific gene}
\usage{
g_per(data_per, gene_pos)
}
\arguments{
\item{data_per}{A data frame containing expression data.}

\item{gene_pos}{Integer index of the gene to permute.}
}
\value{
The data frame with the selected gene permuted.
}
\description{
Permute expression values for a specific gene
}
