% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{gem2mem}
\alias{gem2mem}
\title{Summarize a gene-expression matrix}
\usage{
gem2mem(gem = NULL, meta = NULL, FUN = c("mean", "median"))
}
\arguments{
\item{gem}{A data frame or matrix of single-cell expression values.}

\item{meta}{A character or factor vector of cell-type labels.}

\item{FUN}{Character string specifying \code{"mean"} or \code{"median"}.}
}
\value{
A data frame of summarized expression values.
}
\description{
Summarize a single-cell expression matrix into a cell-type expression matrix.
}
