% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{genBList}
\alias{genBList}
\title{Generate blacklist}
\usage{
genBList(ug, nodes, root = NULL, blackList = NULL)
}
\arguments{
\item{ug}{An undirected graph represented as a data frame.}

\item{nodes}{Character vector of node names.}

\item{root}{Character vector of root nodes.}

\item{blackList}{An optional data frame or character vector of excluded arcs.}
}
\value{
A data frame containing excluded arcs.
}
\description{
Generate blacklist
}
