% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectMat_func.R
\name{get_diffCoeff}
\alias{get_diffCoeff}
\title{Calculate coefficient differences}
\usage{
get_diffCoeff(scm.inter = NULL, scm.ref = NULL, mode = "mean")
}
\arguments{
\item{scm.inter}{A data frame of perturbed \code{bn.fit} coefficients.}

\item{scm.ref}{A data frame of reference \code{bn.fit} coefficients.}

\item{mode}{Character string specifying how coefficient differences are
calculated. Currently, only \code{"mean"} is supported.}
}
\value{
A data frame of coefficient differences.
}
\description{
Calculate coefficient differences from a set of \code{bn.fit} results and an
unperturbed reference.
}
