% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectMat_func.R
\name{get_scm}
\alias{get_scm}
\title{Calculate structural causal model coefficients}
\usage{
get_scm(mem = NULL, graph = NULL, id = NULL)
}
\arguments{
\item{mem}{A data frame of cell-type expression values.}

\item{graph}{A \code{bn.fit} network structure.}

\item{id}{Character identifier for the sampling and perturbation result.}
}
\value{
A data frame of fitted edge coefficients.
}
\description{
Calculate \code{bn.fit} coefficients for a network and a cell-type expression
data frame.
}
