% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{getCoef}
\alias{getCoef}
\title{Extract bn coefficients}
\usage{
getCoef(dag, data)
}
\arguments{
\item{dag}{A \code{bn} object.}

\item{data}{A data frame or matrix used to fit the network.}
}
\value{
A data frame containing arcs and their absolute coefficients.
}
\description{
Extract bn coefficients
}
