% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{getNEdges}
\alias{getNEdges}
\title{Get N edges}
\usage{
getNEdges(dags)
}
\arguments{
\item{dags}{A list of network structures.}
}
\value{
A named numeric vector containing the number of arcs in each
network.
}
\description{
Get N edges
}
