% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{hasCyc}
\alias{hasCyc}
\title{Test a network for cycles}
\usage{
hasCyc(dS)
}
\arguments{
\item{dS}{A summarized network structure.}
}
\value{
\code{TRUE} if the network contains a cycle; otherwise, \code{FALSE}.
}
\description{
Test a network for cycles
}
\keyword{internal}
