% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HMMA}
\alias{HMMA}
\title{HMMA}
\format{
An object of class \code{data.frame} with 21678 rows and 11 columns.
}
\usage{
HMMA
}
\description{
A test microarray expression matrix for HSC differentiation, containing 11
cell types and two lineages.
}
\keyword{datasets}
