% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{learnDAG}
\alias{learnDAG}
\title{Learn the principal DAG from data}
\usage{
learnDAG(
  dat,
  frac = 0.3,
  N_smpl = 10,
  params = c(1:100)/100,
  whiteList = NULL,
  blackList = NULL,
  root = NULL,
  ugMethod = "cmi2ni",
  dagMethod = "hc",
  ncores = 1,
  mode = "single_cell",
  Emin = 1,
  Emax = 50
)
}
\arguments{
\item{dat}{A Seurat object in single-cell mode or an expression matrix in
bulk mode.}

\item{frac}{Fraction of cells included in each sample.}

\item{N_smpl}{Number of cell samples.}

\item{params}{Numeric vector of tuning parameters for the undirected-graph
method.}

\item{whiteList}{A data frame or character vector specifying arcs that must
be present in the network.}

\item{blackList}{A data frame or character vector specifying arcs that must
not be present in the network.}

\item{root}{Character vector of root nodes, which cannot have incoming arcs.}

\item{ugMethod}{Character string specifying the undirected-graph method.}

\item{dagMethod}{Character string specifying the method used to determine arc
directions.}

\item{ncores}{Number of cores used for computation.}

\item{mode}{Character string specifying \code{"single_cell"} or \code{"bulk"} mode.}

\item{Emin}{Minimum number of arcs required for a DAG to be retained.}

\item{Emax}{Maximum number of arcs allowed for a DAG to be retained.}
}
\value{
A \code{bn} object representing the principal directed acyclic graph.
}
\description{
\code{learnDAG} uses \code{BNLearning} to generate a set of possible DAGs representing
the differentiation network and analyses these graphs to give the final DAG.
}
