% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{mat2df}
\alias{mat2df}
\title{Convert a network matrix to a data frame}
\usage{
mat2df(DAG.dm)
}
\arguments{
\item{DAG.dm}{A matrix representing a network structure.}
}
\value{
A data frame containing the summarized network structure.
}
\description{
Transform an adjacency matrix into a summarized network structure.
}
