% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneck_methods.R
\name{MI2}
\alias{MI2}
\title{Calculate a symmetric conditional association score}
\usage{
MI2(x, y, z)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}

\item{z}{A numeric conditioning variable or matrix.}
}
\value{
A numeric symmetric conditional association score.
}
\description{
Calculate a symmetric conditional association score
}
\keyword{internal}
