% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneck_methods.R
\name{network.cmi2ni}
\alias{network.cmi2ni}
\title{Infer a network using CMI2NI}
\usage{
network.cmi2ni(expr.data, lambda)
}
\arguments{
\item{expr.data}{A numeric expression matrix with observations in rows and
genes in columns.}

\item{lambda}{Positive conditional-mutual-information threshold.}
}
\value{
A data frame containing inferred edges and their strengths.
}
\description{
Infer a network using CMI2NI
}
\keyword{internal}
