% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneck_methods.R
\name{network.space}
\alias{network.space}
\title{Infer a graph using the space package}
\usage{
network.space(expr.data, alpha)
}
\arguments{
\item{expr.data}{A numeric expression matrix with observations in rows and
genes in columns.}

\item{alpha}{Positive tuning parameter.}
}
\value{
A data frame containing inferred edges and partial correlations.
}
\description{
Infer a graph using the space package
}
