% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{norm_mat}
\alias{norm_mat}
\title{Reduce undirected edges and normalize edge weight}
\usage{
norm_mat(N)
}
\arguments{
\item{N}{A numeric adjacency matrix.}
}
\value{
A normalized adjacency matrix.
}
\description{
Reduce undirected edges and normalize edge weight
}
