% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{parse_edge_list}
\alias{parse_edge_list}
\title{Parse a whitelist or blacklist}
\usage{
parse_edge_list(edges)
}
\arguments{
\item{edges}{A data frame or character vector of edges.}
}
\value{
A data frame with \code{from} and \code{to} columns, or \code{NULL}.
}
\description{
Parse a whitelist or blacklist
}
\keyword{internal}
