% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_func.R
\name{plot_brweight}
\alias{plot_brweight}
\title{Plot branch weights}
\usage{
plot_brweight(brweight, title = NULL, font_size = 40)
}
\arguments{
\item{brweight}{A list containing branch and weight information.}

\item{title}{Optional character string used as the plot title.}

\item{font_size}{Numeric font size.}
}
\value{
The plotted graph, invisibly.
}
\description{
Plot branch weights
}
