% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{rmCyc}
\alias{rmCyc}
\title{Remove cycles}
\usage{
rmCyc(dS, method = "sorted")
}
\arguments{
\item{dS}{A summarized network structure.}

\item{method}{Character string specifying the cycle-removal method.}
}
\value{
A summarized network structure with cycles removed.
}
\description{
Remove cycles according to the hierarchical structure and network matrix.
}
