% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{rmCyc_sorted}
\alias{rmCyc_sorted}
\title{Remove cycles in descending edge-weight order}
\usage{
rmCyc_sorted(dS)
}
\arguments{
\item{dS}{A summarized network structure.}
}
\value{
A summarized acyclic network structure.
}
\description{
Remove cycles in descending edge-weight order
}
\keyword{internal}
