% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{rmCycL}
\alias{rmCycL}
\title{Remove cycles using graph layout}
\usage{
rmCycL(dS)
}
\arguments{
\item{dS}{A summarized network structure.}
}
\value{
A summarized network structure with cycles removed.
}
\description{
Remove cycles from a network according to its hierarchical layout.
}
