% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{trimDAG}
\alias{trimDAG}
\title{Trim a DAG using fitted coefficients}
\usage{
trimDAG(
  dat_tmp,
  e,
  min_arc = 2,
  max_arc = 4,
  threshold_value = 0.9,
  plot = TRUE
)
}
\arguments{
\item{dat_tmp}{A matrix of gene expression values across cell types.}

\item{e}{A \code{bn.fit} network structure.}

\item{min_arc}{Minimum number of arcs pointing to a node.}

\item{max_arc}{Maximum number of arcs pointing to a node.}

\item{threshold_value}{Numeric threshold used to filter arcs according to
fitted coefficients.}

\item{plot}{Logical; whether to plot the trimmed DAG.}
}
\value{
The filtered network structure.
}
\description{
Modify a network structure according to fitted \code{bn.fit} coefficients.
}
