% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{bootstrap_index}
\alias{bootstrap_index}
\title{Generate bootstrap indices}
\usage{
bootstrap_index(meta, bootstrap_times, ratio)
}
\arguments{
\item{meta}{A character or factor vector of cell-type labels.}

\item{bootstrap_times}{Number of bootstrap samples.}

\item{ratio}{Fraction of cells included in each sample.}
}
\value{
A list of sampling-index vectors.
}
\description{
Create sampling indices for \code{run_diffCoeff()}.
}
