% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{bootstrap_index}
\alias{bootstrap_index}
\title{bootstrap_index}
\usage{
bootstrap_index(meta, bootstrap_times, ratio)
}
\arguments{
\item{meta}{character, the cell type information of single cell data}

\item{bootstrap_times}{numeric, times of sampling, should be some with "bootstrap_time" in "run_diffCoeff"}

\item{ratio}{double, the fraction of cell that are involved in every sample}
}
\value{
list of index for diffCoeff
}
\description{
bootstrap_index: create sampling index for run_diffCoeff
}
