% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_func.R
\name{colorEdges}
\alias{colorEdges}
\title{colorEdges}
\usage{
colorEdges(net, ref, main = "", node_col = NULL)
}
\arguments{
\item{net}{bn, network to be plotted}

\item{ref}{bn, network of reference}

\item{main}{string, title of plot}
}
\value{
plot
}
\description{
colorEdges
}
