% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{df2bn}
\alias{df2bn}
\title{Transform graph structure from data.frame to bn}
\usage{
df2bn(dS, node_names, plot = TRUE)
}
\arguments{
\item{dS}{data.frame, with columns named \code{from} and \code{to}}

\item{node_names}{character}
}
\description{
Transform graph structure from data.frame to bn
}
