% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{df2mat}
\alias{df2mat}
\title{df2mat}
\usage{
df2mat(DAG, ctypes = NULL)
}
\arguments{
\item{DAG}{Summarized net structure}

\item{ctypes}{character, the cell typs used}
}
\value{
A matrix that is corresponding to the summarized net structure
}
\description{
df2mat: transform a summarized net structure to matrix
}
