% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{DG_smpl}
\alias{DG_smpl}
\title{Learn directed graphs from single-cell samples}
\usage{
DG_smpl(
  dat,
  resampling_fraction = 0.3,
  N_smpl = 10,
  params = c(1:100)/100,
  whiteList = NULL,
  blackList = NULL,
  root = NULL,
  ugMethod = "cmi2ni",
  dagMethod = "hc",
  ncores = 64,
  seed = Sys.time()
)
}
\arguments{
\item{dat}{A Seurat object.}

\item{resampling_fraction}{Fraction of cells included in each resampled
dataset.}

\item{N_smpl}{Number of cell samples.}

\item{params}{Numeric vector of tuning parameters for the undirected-graph
method.}

\item{whiteList}{A data frame or character vector specifying arcs that must
be present in the network.}

\item{blackList}{A data frame or character vector specifying arcs that must
not be present in the network.}

\item{root}{Character vector of root nodes, which cannot have incoming arcs.}

\item{ugMethod}{Character string specifying the undirected-graph method.}

\item{dagMethod}{Character string specifying the method used to determine arc
directions.}

\item{ncores}{Number of cores used for computation.}

\item{seed}{Numeric random seed.}
}
\value{
A list of network structures for each parameter value and cell
sample.
}
\description{
Learn network structures across parameter values and cell samples from a
Seurat object.
}
