% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{DG_smpl}
\alias{DG_smpl}
\title{DG_smpl}
\usage{
DG_smpl(
  dat,
  frac = 0.3,
  N_smpl = 10,
  params = c(1:100)/100,
  whiteList = NULL,
  blackList = NULL,
  root = NULL,
  ugMethod = "cmi2ni",
  dagMethod = "hc",
  ncores = 64,
  seed = Sys.time()
)
}
\arguments{
\item{dat}{Seurat object for input}

\item{frac}{numeric, the fraction of cells for every sampling}

\item{N_smpl}{numeric, the total times of sampling}

\item{params}{numeric, the parameters used for UG_method}

\item{whiteList}{data.frame, the list of arcs that must exist in the network}

\item{blackList}{data.frame, the list of arcs that can't exist in the network}

\item{root}{character, the root of the graph, which can't have any node pointing to.}

\item{ugMethod}{character, determine the UG_method used, usually cmi2ni. Other alternatives including ns/GeneNet/glasso/pcacmi/bayesianglasso}

\item{dagMethod}{character, determine the method that calculate the arc direction, usually hc}

\item{ncores}{integer,the cores that used to speed up the calculation}

\item{seed}{numeric}
}
\value{
A list of net structures of different parameters and different sampling
}
\description{
DG_smpl is the function that considers all the parameters in "params" and cell samplings to get a group of net structures based on ugMethod and dagMethod. Compared with DG_grid, DG_smpl is suitable for single cell data in SeuratObject format
}
