% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findFeatures_func.R
\name{diffScore}
\alias{diffScore}
\title{Get diffScore}
\usage{
diffScore(diffCoeff, edgeSet, abs = TRUE)
}
\arguments{
\item{diffCoeff}{data frame, with diffCoeff data previously calculated}

\item{edgeSet}{list of edge names}

\item{abs}{logical, set to TRUE when diffScore is considered the sum of the absolute values of diffCoeff}
}
\value{
list of diffScores
}
\description{
Get diffScore
}
