% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectMat_func.R
\name{EffectMatrix}
\alias{EffectMatrix}
\title{Calculate effect matrix}
\usage{
EffectMatrix(diffBN, mode = "mean")
}
\arguments{
\item{diffBN}{list of diffBN raw result}

\item{mode}{character, determine the diffBN function}
}
\value{
Effect matrix
}
\description{
Calculate effect matrix from perturbation results.
}
