% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectMat_func.R
\name{g_per}
\alias{g_per}
\title{Permutate specific gene data}
\usage{
g_per(data_per, gene_pos)
}
\arguments{
\item{data_per}{data.frame}

\item{gene_pos}{interger}
}
\value{
data.frame
}
\description{
Permutate specific gene data
}
