% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{gem2mem}
\alias{gem2mem}
\title{gem2mem}
\usage{
gem2mem(gem = NULL, meta = NULL, FUN = c("mean", "median"))
}
\arguments{
\item{gem}{data.frame of single cell expression}

\item{meta}{character, cell type information of single cell expression data}

\item{FUN}{character, can be 'mean' or 'median'}
}
\value{
data.frame
}
\description{
gem2mem is a function that can summary a single cell expression matrix to a cell type expression matrix using the metadata and method given
Usually, FUN='mean'
}
