% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectMat_func.R
\name{get_diffCoeff}
\alias{get_diffCoeff}
\title{get_diffCoeff}
\usage{
get_diffCoeff(scm.inter = NULL, scm.ref = NULL, mode = "mean")
}
\arguments{
\item{scm.inter}{data.frame of bn.fit results}

\item{scm.ref}{data.frame of bn.fit reference, which don't have any gene deletion or permutation}

\item{mode}{character, determine how the diffCoeff is calculate, usually use "mean"}
}
\value{
data.frame of diffCoeff
}
\description{
get_diffCoeff: based on a set of bn.fit results and a bn.fir reference, calculate the diffCoeff results
}
