% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectMat_func.R
\name{get_scm}
\alias{get_scm}
\title{get_scm}
\usage{
get_scm(mem = NULL, graph = NULL, id = NULL)
}
\arguments{
\item{mem}{data.frame of cell type expression}

\item{graph}{bn.fit structure}

\item{id}{character, id should be given to label the information of sampling and permutation, best in the form of s1p1}
}
\value{
bn.fit results
}
\description{
get_scm:calculate bn.fit results according to a certain neetwork structure and a cell type expression data.frame
}
