% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectMat_func.R
\name{GRNPerturbResult}
\alias{GRNPerturbResult}
\title{CellOracle-based GRN perturbation}
\usage{
GRNPerturbResult(
  object,
  net_struc,
  oracle,
  links,
  save_dir,
  group.by = "celltype",
  assay = "RNA",
  layer = "data",
  genes = NULL,
  perturb_value = 0,
  n_propagation = 3,
  alpha = 10,
  filter_links = TRUE,
  overwrite = FALSE,
  verbal = TRUE
)
}
\arguments{
\item{object}{Seurat object}

\item{net_struc}{Cell-type DAG}

\item{oracle}{Preprocessed CellOracle Oracle object}

\item{links}{Path to a CellOracle Links file}

\item{save_dir}{Directory for intermediate SCM results}

\item{group.by}{Cell-type column in Seurat metadata}

\item{assay}{Assay used for SCM calculation}

\item{layer}{Expression layer used for SCM calculation}

\item{genes}{Genes to perturb; NULL uses all available regulatory genes}

\item{perturb_value}{Target expression value}

\item{n_propagation}{Number of CellOracle propagation steps}

\item{alpha}{Ridge coefficient used for GRN fitting}

\item{filter_links}{Whether to filter Links before GRN fitting}

\item{overwrite}{Whether to overwrite cached results}

\item{verbal}{Whether to print progress}
}
\value{
A PerturbResult-compatible list
}
\description{
CellOracle-based GRN perturbation
}
