% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HMR}
\alias{HMR}
\title{HMR}
\format{
An object of class \code{Seurat} with 40198 rows and 716 columns.
}
\usage{
HMR
}
\description{
A Seurat object:
Single cell data of HMR
}
\keyword{datasets}
