% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneck_methods.R
\name{network.cmi2ni}
\alias{network.cmi2ni}
\title{Use cmi2ni to predict graph structure}
\usage{
network.cmi2ni(expr.data, lambda)
}
\arguments{
\item{expr.data}{data.frame}

\item{lambda}{double}
}
\value{
data.frame
}
\description{
Use cmi2ni to predict graph structure
}
