% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectMat_func.R
\name{PerturbResult}
\alias{PerturbResult}
\title{PerturbResult}
\usage{
PerturbResult(
  net_struc,
  data,
  meta = NULL,
  index = NULL,
  n_sample = 1,
  n_permutation = 1,
  deletion = T,
  mode = "single_cell",
  ncores = 1,
  verbal = F,
  replace = F,
  oracle = NULL,
  links = NULL,
  save_dir = NULL,
  group.by = "celltype",
  assay = "RNA",
  layer = "data",
  genes = NULL,
  perturb_value = 0,
  n_propagation = 3,
  alpha = 10,
  filter_links = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{net_struc}{bn.fit structure}

\item{data}{data.frame of the expression data}

\item{meta}{character of metadata corresponding to expression data}

\item{index}{list of sampling results}

\item{n_sample}{integer, times of cell sampling, usually 20-50}

\item{n_permutation}{integer, times of gene permutation, usually 20-50}

\item{deletion, }{logical, whether the permutation uses deletion, if this is set to true, n_permutation and replace will not be used}

\item{mode}{character, can be 'single_cell' or 'bulk'}

\item{ncores}{integer, sets the number of cores used in the parallel computing}

\item{verbal}{logical}

\item{replace}{logical, whether the sampling will be performed with replacing}
}
\value{
diffCoeff results
}
\description{
PerturbResult is a function that calculate coefficients of the graph structure based on a given expression data
}
