% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_func.R
\name{plot_brweight}
\alias{plot_brweight}
\title{plot_brweight}
\usage{
plot_brweight(brweight, title = NULL, font_size = 40)
}
\arguments{
\item{brweight}{numeric}

\item{title}{character}

\item{font_size}{interger}
}
\value{
plot
}
\description{
plot_brweight
}
