% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R
\name{rmCyc2}
\alias{rmCyc2}
\title{Reduce undirected edges to directed edges}
\usage{
rmCyc2(dS)
}
\arguments{
\item{dS}{list}
}
\value{
list
}
\description{
Reduce undirected edges to directed edges
}
