% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectMat_func.R
\name{run_diffCoeff}
\alias{run_diffCoeff}
\title{run_diffCoeff}
\usage{
run_diffCoeff(
  net_struc,
  data,
  meta,
  index,
  n_sample = 20,
  n_permutation = 20,
  diffCoeff_mode = "mean",
  verbal = F,
  replace = F
)
}
\arguments{
\item{net_struc}{bn.fit structure}

\item{data}{data.frame of single cell expression data}

\item{meta}{character of metadata corresponding to expression data}

\item{index}{list of sampling results}

\item{n_sample}{integer, times of cell sampling, usually 20-50}

\item{n_permutation}{integer, times of gene permutation, usually 20-50}

\item{diffCoeff_mode}{character, determine the diffCoeff function, usually "mean"}
}
\value{
diffCoeff results
}
\description{
run_diffCoeff is a function that calculate diffCoeff results based on a single cell expression data
}
