% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNLearning_func.R, R/findFeatures_func.R
\name{setEdges}
\alias{setEdges}
\title{Generate edge set}
\usage{
setEdges(fromSet, toSet, sep = "~")

setEdges(fromSet, toSet, sep = "~")
}
\arguments{
\item{fromSet}{list of starting nodes}

\item{toSet}{list of ending nodes}

\item{sep}{character, used for sepearating names}
}
\value{
character

list of edge names
}
\description{
Generate edge set

Generate edge set
}
\examples{
ctypes <- c("MPPa", "MPPb")
s.diffScore <- setEdges(ctypes, ctypes, sep = "~")
}
