% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_func.R
\name{ugPlot}
\alias{ugPlot}
\title{ugPlot}
\usage{
ugPlot(ug, weight = FALSE, ...)
}
\arguments{
\item{ug}{data.frame}

\item{weight}{logic}

\item{...}{other}
}
\value{
plot
}
\description{
ugPlot
}
